mcp_service/requirements.txt
mcp
sentence-transformers
faiss-cpu
pandas
numpy
uvicorn
starlette
requests
fastapi
pydantic
